decouple coordinator with server
