+++ pkg/schedule/schedulers/range_cluster.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package schedule
+package schedulers
 
 import (
 	"github.com/docker/go-units"
@@ -20,28 +20,28 @@ import (
 	sche "github.com/tikv/pd/pkg/schedule/core"
 )
 
-// RangeCluster isolates the cluster by range.
-type RangeCluster struct {
+// rangeCluster isolates the cluster by range.
+type rangeCluster struct {
 	sche.ClusterInformer
 	subCluster        *core.BasicCluster // Collect all regions belong to the range.
 	tolerantSizeRatio float64
 }
 
-// GenRangeCluster gets a range cluster by specifying start key and end key.
+// genRangeCluster gets a range cluster by specifying start key and end key.
 // The cluster can only know the regions within [startKey, endKey].
-func GenRangeCluster(cluster sche.ClusterInformer, startKey, endKey []byte) *RangeCluster {
+func genRangeCluster(cluster sche.ClusterInformer, startKey, endKey []byte) *rangeCluster {
 	subCluster := core.NewBasicCluster()
 	for _, r := range cluster.ScanRegions(startKey, endKey, -1) {
 		origin, overlaps, rangeChanged := subCluster.SetRegion(r)
 		subCluster.UpdateSubTree(r, origin, overlaps, rangeChanged)
 	}
-	return &RangeCluster{
+	return &rangeCluster{
 		ClusterInformer: cluster,
 		subCluster:      subCluster,
 	}
 }
 
-func (r *RangeCluster) updateStoreInfo(s *core.StoreInfo) *core.StoreInfo {
+func (r *rangeCluster) updateStoreInfo(s *core.StoreInfo) *core.StoreInfo {
 	id := s.GetID()
 
 	used := float64(s.GetUsedSize()) / units.MiB
@@ -69,7 +69,7 @@ func (r *RangeCluster) updateStoreInfo(s *core.StoreInfo) *core.StoreInfo {
 }
 
 // GetStore searches for a store by ID.
-func (r *RangeCluster) GetStore(id uint64) *core.StoreInfo {
+func (r *rangeCluster) GetStore(id uint64) *core.StoreInfo {
 	s := r.ClusterInformer.GetStore(id)
 	if s == nil {
 		return nil
@@ -78,7 +78,7 @@ func (r *RangeCluster) GetStore(id uint64) *core.StoreInfo {
 }
 
 // GetStores returns all Stores in the cluster.
-func (r *RangeCluster) GetStores() []*core.StoreInfo {
+func (r *rangeCluster) GetStores() []*core.StoreInfo {
 	stores := r.ClusterInformer.GetStores()
 	newStores := make([]*core.StoreInfo, 0, len(stores))
 	for _, s := range stores {
@@ -88,12 +88,12 @@ func (r *RangeCluster) GetStores() []*core.StoreInfo {
 }
 
 // SetTolerantSizeRatio sets the tolerant size ratio.
-func (r *RangeCluster) SetTolerantSizeRatio(ratio float64) {
+func (r *rangeCluster) SetTolerantSizeRatio(ratio float64) {
 	r.tolerantSizeRatio = ratio
 }
 
 // GetTolerantSizeRatio gets the tolerant size ratio.
-func (r *RangeCluster) GetTolerantSizeRatio() float64 {
+func (r *rangeCluster) GetTolerantSizeRatio() float64 {
 	if r.tolerantSizeRatio != 0 {
 		return r.tolerantSizeRatio
 	}
@@ -101,22 +101,22 @@ func (r *RangeCluster) GetTolerantSizeRatio() float64 {
 }
 
 // RandFollowerRegions returns a random region that has a follower on the store.
-func (r *RangeCluster) RandFollowerRegions(storeID uint64, ranges []core.KeyRange) []*core.RegionInfo {
+func (r *rangeCluster) RandFollowerRegions(storeID uint64, ranges []core.KeyRange) []*core.RegionInfo {
 	return r.subCluster.RandFollowerRegions(storeID, ranges)
 }
 
 // RandLeaderRegions returns a random region that has leader on the store.
-func (r *RangeCluster) RandLeaderRegions(storeID uint64, ranges []core.KeyRange) []*core.RegionInfo {
+func (r *rangeCluster) RandLeaderRegions(storeID uint64, ranges []core.KeyRange) []*core.RegionInfo {
 	return r.subCluster.RandLeaderRegions(storeID, ranges)
 }
 
 // GetAverageRegionSize returns the average region approximate size.
-func (r *RangeCluster) GetAverageRegionSize() int64 {
+func (r *rangeCluster) GetAverageRegionSize() int64 {
 	return r.subCluster.GetAverageRegionSize()
 }
 
 // GetRegionStores returns all stores that contains the region's peer.
-func (r *RangeCluster) GetRegionStores(region *core.RegionInfo) []*core.StoreInfo {
+func (r *rangeCluster) GetRegionStores(region *core.RegionInfo) []*core.StoreInfo {
 	stores := r.ClusterInformer.GetRegionStores(region)
 	newStores := make([]*core.StoreInfo, 0, len(stores))
 	for _, s := range stores {
@@ -126,7 +126,7 @@ func (r *RangeCluster) GetRegionStores(region *core.RegionInfo) []*core.StoreInf
 }
 
 // GetFollowerStores returns all stores that contains the region's follower peer.
-func (r *RangeCluster) GetFollowerStores(region *core.RegionInfo) []*core.StoreInfo {
+func (r *rangeCluster) GetFollowerStores(region *core.RegionInfo) []*core.StoreInfo {
 	stores := r.ClusterInformer.GetFollowerStores(region)
 	newStores := make([]*core.StoreInfo, 0, len(stores))
 	for _, s := range stores {
@@ -136,7 +136,7 @@ func (r *RangeCluster) GetFollowerStores(region *core.RegionInfo) []*core.StoreI
 }
 
 // GetLeaderStore returns all stores that contains the region's leader peer.
-func (r *RangeCluster) GetLeaderStore(region *core.RegionInfo) *core.StoreInfo {
+func (r *rangeCluster) GetLeaderStore(region *core.RegionInfo) *core.StoreInfo {
 	s := r.ClusterInformer.GetLeaderStore(region)
 	if s != nil {
 		return r.updateStoreInfo(s)
