+++ server/api/diagnostic_test.go
@@ -23,11 +23,11 @@ import (
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/stretchr/testify/suite"
 	"github.com/tikv/pd/pkg/core"
+	"github.com/tikv/pd/pkg/schedule"
 	"github.com/tikv/pd/pkg/schedule/schedulers"
 	"github.com/tikv/pd/pkg/utils/apiutil"
 	tu "github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/server"
-	"github.com/tikv/pd/server/cluster"
 	"github.com/tikv/pd/server/config"
 )
 
@@ -66,7 +66,7 @@ func (suite *diagnosticTestSuite) TearDownSuite() {
 func (suite *diagnosticTestSuite) checkStatus(status string, url string) {
 	re := suite.Require()
 	suite.Eventually(func() bool {
-		result := &cluster.DiagnosticResult{}
+		result := &schedule.DiagnosticResult{}
 		err := tu.ReadGetJSON(re, testDialClient, url, result)
 		suite.NoError(err)
 		return result.Status == status
@@ -95,7 +95,7 @@ func (suite *diagnosticTestSuite) TestSchedulerDiagnosticAPI() {
 	suite.True(cfg.Schedule.EnableDiagnostic)
 
 	balanceRegionURL := suite.urlPrefix + "/" + schedulers.BalanceRegionName
-	result := &cluster.DiagnosticResult{}
+	result := &schedule.DiagnosticResult{}
 	err = tu.ReadGetJSON(re, testDialClient, balanceRegionURL, result)
 	suite.NoError(err)
 	suite.Equal("disabled", result.Status)
