+++ server/cluster/cluster_test.go
@@ -16,6 +16,7 @@ package cluster
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"math"
 	"math/rand"
@@ -23,22 +24,33 @@ import (
 	"testing"
 	"time"
 
+	"github.com/docker/go-units"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
+	"github.com/pingcap/kvproto/pkg/eraftpb"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/pd/pkg/core"
+	"github.com/tikv/pd/pkg/core/constant"
+	"github.com/tikv/pd/pkg/core/storelimit"
 	"github.com/tikv/pd/pkg/errs"
 	"github.com/tikv/pd/pkg/id"
+	"github.com/tikv/pd/pkg/mock/mockhbstream"
 	"github.com/tikv/pd/pkg/mock/mockid"
 	"github.com/tikv/pd/pkg/progress"
+	"github.com/tikv/pd/pkg/schedule"
+	sche "github.com/tikv/pd/pkg/schedule/core"
 	"github.com/tikv/pd/pkg/schedule/filter"
+	"github.com/tikv/pd/pkg/schedule/hbstream"
 	"github.com/tikv/pd/pkg/schedule/labeler"
+	"github.com/tikv/pd/pkg/schedule/operator"
 	"github.com/tikv/pd/pkg/schedule/placement"
 	"github.com/tikv/pd/pkg/schedule/schedulers"
 	"github.com/tikv/pd/pkg/statistics"
 	"github.com/tikv/pd/pkg/storage"
+	"github.com/tikv/pd/pkg/utils/operatorutil"
+	"github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/pkg/utils/typeutil"
 	"github.com/tikv/pd/pkg/versioninfo"
 	"github.com/tikv/pd/server/config"
@@ -224,7 +236,7 @@ func TestSetOfflineStore(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.ruleManager = placement.NewRuleManager(storage.NewStorageWithMemoryBackend(), cluster, cluster.GetOpts())
 	if opt.IsPlacementRulesEnabled() {
 		err := cluster.ruleManager.Initialize(opt.GetMaxReplicas(), opt.GetLocationLabels())
@@ -290,7 +302,7 @@ func TestSetOfflineWithReplica(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	// Put 4 stores.
 	for _, store := range newTestStores(4, "2.0.0") {
@@ -329,7 +341,7 @@ func TestSetOfflineStoreWithEvictLeader(t *testing.T) {
 	re.NoError(err)
 	opt.SetMaxReplicas(1)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	// Put 3 stores.
 	for _, store := range newTestStores(3, "2.0.0") {
@@ -375,7 +387,7 @@ func TestReuseAddress(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	// Put 4 stores.
 	for _, store := range newTestStores(4, "2.0.0") {
 		re.NoError(cluster.PutStore(store.GetMeta()))
@@ -421,7 +433,7 @@ func TestUpStore(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.ruleManager = placement.NewRuleManager(storage.NewStorageWithMemoryBackend(), cluster, cluster.GetOpts())
 	if opt.IsPlacementRulesEnabled() {
 		err := cluster.ruleManager.Initialize(opt.GetMaxReplicas(), opt.GetLocationLabels())
@@ -466,7 +478,7 @@ func TestRemovingProcess(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.SetPrepared()
 
 	// Put 5 stores.
@@ -524,7 +536,7 @@ func TestDeleteStoreUpdatesClusterVersion(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.ruleManager = placement.NewRuleManager(storage.NewStorageWithMemoryBackend(), cluster, cluster.GetOpts())
 	if opt.IsPlacementRulesEnabled() {
 		err := cluster.ruleManager.Initialize(opt.GetMaxReplicas(), opt.GetLocationLabels())
@@ -584,7 +596,7 @@ func TestRegionHeartbeatHotStat(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	newTestStores(4, "2.0.0")
 	peers := []*metapb.Peer{
 		{
@@ -646,7 +658,7 @@ func TestBucketHeartbeat(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	// case1: region is not exist
 	buckets := &metapb.Buckets{
@@ -705,7 +717,7 @@ func TestRegionHeartbeat(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	n, np := uint64(3), uint64(3)
 	cluster.wg.Add(1)
 	go cluster.runUpdateStoreStats()
@@ -940,7 +952,7 @@ func TestRegionFlowChanged(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	regions := []*core.RegionInfo{core.NewTestRegionInfo(1, 1, []byte{}, []byte{})}
 	processRegions := func(regions []*core.RegionInfo) {
 		for _, r := range regions {
@@ -965,7 +977,7 @@ func TestRegionSizeChanged(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.regionStats = statistics.NewRegionStatistics(cluster.GetOpts(), cluster.ruleManager, cluster.storeConfigManager)
 	region := newTestRegions(1, 3, 3)[0]
 	cluster.opt.GetMaxMergeRegionKeys()
@@ -1008,7 +1020,7 @@ func TestConcurrentReportBucket(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	regions := []*core.RegionInfo{core.NewTestRegionInfo(1, 1, []byte{}, []byte{})}
 	heartbeatRegions(re, cluster, regions)
@@ -1038,7 +1050,7 @@ func TestConcurrentRegionHeartbeat(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	regions := []*core.RegionInfo{core.NewTestRegionInfo(1, 1, []byte{}, []byte{})}
 	regions = core.SplitRegions(regions)
@@ -1116,7 +1128,7 @@ func TestRegionLabelIsolationLevel(t *testing.T) {
 	r := core.NewRegionInfo(region, peers[0])
 	re.NoError(cluster.putRegion(r))
 
-	cluster.updateRegionsLabelLevelStats([]*core.RegionInfo{r})
+	cluster.UpdateRegionsLabelLevelStats([]*core.RegionInfo{r})
 	counter := cluster.labelLevelStats.GetLabelCounter()
 	re.Equal(0, counter["none"])
 	re.Equal(1, counter["zone"])
@@ -1158,7 +1170,7 @@ func TestHeartbeatSplit(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	// 1: [nil, nil)
 	region1 := core.NewRegionInfo(&metapb.Region{Id: 1, RegionEpoch: &metapb.RegionEpoch{Version: 1, ConfVer: 1}}, nil)
@@ -1202,7 +1214,7 @@ func TestRegionSplitAndMerge(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	regions := []*core.RegionInfo{core.NewTestRegionInfo(1, 1, []byte{}, []byte{})}
 
@@ -1240,7 +1252,7 @@ func TestOfflineAndMerge(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.ruleManager = placement.NewRuleManager(storage.NewStorageWithMemoryBackend(), cluster, cluster.GetOpts())
 	if opt.IsPlacementRulesEnabled() {
 		err := cluster.ruleManager.Initialize(opt.GetMaxReplicas(), opt.GetLocationLabels())
@@ -1249,7 +1261,7 @@ func TestOfflineAndMerge(t *testing.T) {
 		}
 	}
 	cluster.regionStats = statistics.NewRegionStatistics(cluster.GetOpts(), cluster.ruleManager, cluster.storeConfigManager)
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 
 	// Put 4 stores.
 	for _, store := range newTestStores(4, "5.0.0") {
@@ -1354,7 +1366,7 @@ func TestUpdateStorePendingPeerCount(t *testing.T) {
 	_, opt, err := newTestScheduleConfig()
 	re.NoError(err)
 	tc := newTestCluster(ctx, opt)
-	tc.RaftCluster.coordinator = newCoordinator(ctx, tc.RaftCluster, nil)
+	tc.RaftCluster.coordinator = schedule.NewCoordinator(ctx, tc.RaftCluster, nil)
 	stores := newTestStores(5, "2.0.0")
 	for _, s := range stores {
 		re.NoError(tc.putStoreLocked(s))
@@ -1502,7 +1514,7 @@ func TestCalculateStoreSize1(t *testing.T) {
 	cfg.EnablePlacementRules = true
 	opt.SetReplicationConfig(cfg)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.regionStats = statistics.NewRegionStatistics(cluster.GetOpts(), cluster.ruleManager, cluster.storeConfigManager)
 
 	// Put 10 stores.
@@ -1585,7 +1597,7 @@ func TestCalculateStoreSize2(t *testing.T) {
 	opt.SetReplicationConfig(cfg)
 	opt.SetMaxReplicas(3)
 	cluster := newTestRaftCluster(ctx, mockid.NewIDAllocator(), opt, storage.NewStorageWithMemoryBackend(), core.NewBasicCluster())
-	cluster.coordinator = newCoordinator(ctx, cluster, nil)
+	cluster.coordinator = schedule.NewCoordinator(ctx, cluster, nil)
 	cluster.regionStats = statistics.NewRegionStatistics(cluster.GetOpts(), cluster.ruleManager, cluster.storeConfigManager)
 
 	// Put 10 stores.
@@ -2138,4 +2150,1376 @@ func checkStaleRegion(origin *metapb.Region, region *metapb.Region) error {
 	}
 
 	return nil
+}
+
+func newTestOperator(regionID uint64, regionEpoch *metapb.RegionEpoch, kind operator.OpKind, steps ...operator.OpStep) *operator.Operator {
+	return operator.NewTestOperator(regionID, regionEpoch, kind, steps...)
+}
+
+func (c *testCluster) AllocPeer(storeID uint64) (*metapb.Peer, error) {
+	id, err := c.GetAllocator().Alloc()
+	if err != nil {
+		return nil, err
+	}
+	return &metapb.Peer{Id: id, StoreId: storeID}, nil
+}
+
+func (c *testCluster) addRegionStore(storeID uint64, regionCount int, regionSizes ...uint64) error {
+	var regionSize uint64
+	if len(regionSizes) == 0 {
+		regionSize = uint64(regionCount) * 10
+	} else {
+		regionSize = regionSizes[0]
+	}
+
+	stats := &pdpb.StoreStats{}
+	stats.Capacity = 100 * units.GiB
+	stats.UsedSize = regionSize * units.MiB
+	stats.Available = stats.Capacity - stats.UsedSize
+	newStore := core.NewStoreInfo(&metapb.Store{Id: storeID},
+		core.SetStoreStats(stats),
+		core.SetRegionCount(regionCount),
+		core.SetRegionSize(int64(regionSize)),
+		core.SetLastHeartbeatTS(time.Now()),
+	)
+
+	c.SetStoreLimit(storeID, storelimit.AddPeer, 60)
+	c.SetStoreLimit(storeID, storelimit.RemovePeer, 60)
+	c.Lock()
+	defer c.Unlock()
+	return c.putStoreLocked(newStore)
+}
+
+func (c *testCluster) addLeaderRegion(regionID uint64, leaderStoreID uint64, followerStoreIDs ...uint64) error {
+	region := newTestRegionMeta(regionID)
+	leader, _ := c.AllocPeer(leaderStoreID)
+	region.Peers = []*metapb.Peer{leader}
+	for _, followerStoreID := range followerStoreIDs {
+		peer, _ := c.AllocPeer(followerStoreID)
+		region.Peers = append(region.Peers, peer)
+	}
+	regionInfo := core.NewRegionInfo(region, leader, core.SetApproximateSize(10), core.SetApproximateKeys(10))
+	return c.putRegion(regionInfo)
+}
+
+func (c *testCluster) updateLeaderCount(storeID uint64, leaderCount int) error {
+	store := c.GetStore(storeID)
+	newStore := store.Clone(
+		core.SetLeaderCount(leaderCount),
+		core.SetLeaderSize(int64(leaderCount)*10),
+	)
+	c.Lock()
+	defer c.Unlock()
+	return c.putStoreLocked(newStore)
+}
+
+func (c *testCluster) addLeaderStore(storeID uint64, leaderCount int) error {
+	stats := &pdpb.StoreStats{}
+	newStore := core.NewStoreInfo(&metapb.Store{Id: storeID},
+		core.SetStoreStats(stats),
+		core.SetLeaderCount(leaderCount),
+		core.SetLeaderSize(int64(leaderCount)*10),
+		core.SetLastHeartbeatTS(time.Now()),
+	)
+
+	c.SetStoreLimit(storeID, storelimit.AddPeer, 60)
+	c.SetStoreLimit(storeID, storelimit.RemovePeer, 60)
+	c.Lock()
+	defer c.Unlock()
+	return c.putStoreLocked(newStore)
+}
+
+func (c *testCluster) setStoreDown(storeID uint64) error {
+	store := c.GetStore(storeID)
+	newStore := store.Clone(
+		core.UpStore(),
+		core.SetLastHeartbeatTS(typeutil.ZeroTime),
+	)
+	c.Lock()
+	defer c.Unlock()
+	return c.putStoreLocked(newStore)
+}
+
+func (c *testCluster) setStoreOffline(storeID uint64) error {
+	store := c.GetStore(storeID)
+	newStore := store.Clone(core.OfflineStore(false))
+	c.Lock()
+	defer c.Unlock()
+	return c.putStoreLocked(newStore)
+}
+
+func (c *testCluster) LoadRegion(regionID uint64, followerStoreIDs ...uint64) error {
+	//  regions load from etcd will have no leader
+	region := newTestRegionMeta(regionID)
+	region.Peers = []*metapb.Peer{}
+	for _, id := range followerStoreIDs {
+		peer, _ := c.AllocPeer(id)
+		region.Peers = append(region.Peers, peer)
+	}
+	return c.putRegion(core.NewRegionInfo(region, nil))
+}
+
+func TestBasic(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+	oc := co.GetOperatorController()
+
+	re.NoError(tc.addLeaderRegion(1, 1))
+
+	op1 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpLeader)
+	oc.AddWaitingOperator(op1)
+	re.Equal(uint64(1), oc.OperatorCount(operator.OpLeader))
+	re.Equal(op1.RegionID(), oc.GetOperator(1).RegionID())
+
+	// Region 1 already has an operator, cannot add another one.
+	op2 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpRegion)
+	oc.AddWaitingOperator(op2)
+	re.Equal(uint64(0), oc.OperatorCount(operator.OpRegion))
+
+	// Remove the operator manually, then we can add a new operator.
+	re.True(oc.RemoveOperator(op1))
+	op3 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpRegion)
+	oc.AddWaitingOperator(op3)
+	re.Equal(uint64(1), oc.OperatorCount(operator.OpRegion))
+	re.Equal(op3.RegionID(), oc.GetOperator(1).RegionID())
+}
+
+func TestDispatch(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+	co.GetPrepareChecker().SetPrepared()
+	// Transfer peer from store 4 to store 1.
+	re.NoError(tc.addRegionStore(4, 40))
+	re.NoError(tc.addRegionStore(3, 30))
+	re.NoError(tc.addRegionStore(2, 20))
+	re.NoError(tc.addRegionStore(1, 10))
+	re.NoError(tc.addLeaderRegion(1, 2, 3, 4))
+
+	// Transfer leader from store 4 to store 2.
+	re.NoError(tc.updateLeaderCount(4, 50))
+	re.NoError(tc.updateLeaderCount(3, 50))
+	re.NoError(tc.updateLeaderCount(2, 20))
+	re.NoError(tc.updateLeaderCount(1, 10))
+	re.NoError(tc.addLeaderRegion(2, 4, 3, 2))
+
+	go co.RunUntilStop()
+
+	// Wait for schedule and turn off balance.
+	waitOperator(re, co, 1)
+	operatorutil.CheckTransferPeer(re, co.GetOperatorController().GetOperator(1), operator.OpKind(0), 4, 1)
+	re.NoError(co.RemoveScheduler(schedulers.BalanceRegionName))
+	waitOperator(re, co, 2)
+	operatorutil.CheckTransferLeader(re, co.GetOperatorController().GetOperator(2), operator.OpKind(0), 4, 2)
+	re.NoError(co.RemoveScheduler(schedulers.BalanceLeaderName))
+
+	stream := mockhbstream.NewHeartbeatStream()
+
+	// Transfer peer.
+	region := tc.GetRegion(1).Clone()
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitAddLearner(re, stream, region, 1)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitPromoteLearner(re, stream, region, 1)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitRemovePeer(re, stream, region, 4)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+
+	// Transfer leader.
+	region = tc.GetRegion(2).Clone()
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitTransferLeader(re, stream, region, 2)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+}
+
+func dispatchHeartbeat(co *schedule.Coordinator, region *core.RegionInfo, stream hbstream.HeartbeatStream) error {
+	co.GetHeartbeatStreams().BindStream(region.GetLeader().GetStoreId(), stream)
+	if err := co.GetCluster().(*RaftCluster).putRegion(region.Clone()); err != nil {
+		return err
+	}
+	co.GetOperatorController().Dispatch(region, operator.DispatchFromHeartBeat)
+	return nil
+}
+
+func TestCollectMetricsConcurrent(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, func(tc *testCluster) {
+		tc.regionStats = statistics.NewRegionStatistics(tc.GetOpts(), nil, tc.storeConfigManager)
+	}, func(co *schedule.Coordinator) { co.Run() }, re)
+	defer cleanup()
+
+	// Make sure there are no problem when concurrent write and read
+	var wg sync.WaitGroup
+	count := 10
+	wg.Add(count + 1)
+	for i := 0; i <= count; i++ {
+		go func(i int) {
+			defer wg.Done()
+			for j := 0; j < 1000; j++ {
+				re.NoError(tc.addRegionStore(uint64(i%5), rand.Intn(200)))
+			}
+		}(i)
+	}
+	for i := 0; i < 1000; i++ {
+		co.CollectHotSpotMetrics()
+		co.CollectSchedulerMetrics()
+		co.GetCluster().(*RaftCluster).collectClusterMetrics()
+	}
+	co.ResetHotSpotMetrics()
+	co.ResetSchedulerMetrics()
+	co.GetCluster().(*RaftCluster).resetClusterMetrics()
+	wg.Wait()
+}
+
+func TestCollectMetrics(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, func(tc *testCluster) {
+		tc.regionStats = statistics.NewRegionStatistics(tc.GetOpts(), nil, tc.storeConfigManager)
+	}, func(co *schedule.Coordinator) { co.Run() }, re)
+	defer cleanup()
+	count := 10
+	for i := 0; i <= count; i++ {
+		for k := 0; k < 200; k++ {
+			item := &statistics.HotPeerStat{
+				StoreID:   uint64(i % 5),
+				RegionID:  uint64(i*1000 + k),
+				Loads:     []float64{10, 20, 30},
+				HotDegree: 10,
+				AntiCount: statistics.HotRegionAntiCount, // for write
+			}
+			tc.hotStat.HotCache.Update(item, statistics.Write)
+		}
+	}
+	for i := 0; i < 1000; i++ {
+		co.CollectHotSpotMetrics()
+		co.CollectSchedulerMetrics()
+		co.GetCluster().(*RaftCluster).collectClusterMetrics()
+	}
+	stores := co.GetCluster().GetStores()
+	regionStats := co.GetCluster().RegionWriteStats()
+	status1 := statistics.CollectHotPeerInfos(stores, regionStats)
+	status2 := statistics.GetHotStatus(stores, co.GetCluster().GetStoresLoads(), regionStats, statistics.Write, co.GetCluster().GetOpts().IsTraceRegionFlow())
+	for _, s := range status2.AsLeader {
+		s.Stats = nil
+	}
+	for _, s := range status2.AsPeer {
+		s.Stats = nil
+	}
+	re.Equal(status1, status2)
+	co.ResetHotSpotMetrics()
+	co.ResetSchedulerMetrics()
+	co.GetCluster().(*RaftCluster).resetClusterMetrics()
+}
+
+func prepare(setCfg func(*config.ScheduleConfig), setTc func(*testCluster), run func(*schedule.Coordinator), re *require.Assertions) (*testCluster, *schedule.Coordinator, func()) {
+	ctx, cancel := context.WithCancel(context.Background())
+	cfg, opt, err := newTestScheduleConfig()
+	re.NoError(err)
+	if setCfg != nil {
+		setCfg(cfg)
+	}
+	tc := newTestCluster(ctx, opt)
+	hbStreams := hbstream.NewTestHeartbeatStreams(ctx, tc.meta.GetId(), tc, true /* need to run */)
+	if setTc != nil {
+		setTc(tc)
+	}
+	co := schedule.NewCoordinator(ctx, tc.RaftCluster, hbStreams)
+	if run != nil {
+		run(co)
+	}
+	return tc, co, func() {
+		co.Stop()
+		co.GetWaitGroup().Wait()
+		hbStreams.Close()
+		cancel()
+	}
+}
+
+func checkRegionAndOperator(re *require.Assertions, tc *testCluster, co *schedule.Coordinator, regionID uint64, expectAddOperator int) {
+	ops := co.GetCheckerController().CheckRegion(tc.GetRegion(regionID))
+	if ops == nil {
+		re.Equal(0, expectAddOperator)
+	} else {
+		re.Equal(expectAddOperator, co.GetOperatorController().AddWaitingOperator(ops...))
+	}
+}
+
+func TestCheckRegion(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	hbStreams, opt := co.GetHeartbeatStreams(), tc.opt
+	defer cleanup()
+
+	re.NoError(tc.addRegionStore(4, 4))
+	re.NoError(tc.addRegionStore(3, 3))
+	re.NoError(tc.addRegionStore(2, 2))
+	re.NoError(tc.addRegionStore(1, 1))
+	re.NoError(tc.addLeaderRegion(1, 2, 3))
+	checkRegionAndOperator(re, tc, co, 1, 1)
+	operatorutil.CheckAddPeer(re, co.GetOperatorController().GetOperator(1), operator.OpReplica, 1)
+	checkRegionAndOperator(re, tc, co, 1, 0)
+
+	r := tc.GetRegion(1)
+	p := &metapb.Peer{Id: 1, StoreId: 1, Role: metapb.PeerRole_Learner}
+	r = r.Clone(
+		core.WithAddPeer(p),
+		core.WithPendingPeers(append(r.GetPendingPeers(), p)),
+	)
+	re.NoError(tc.putRegion(r))
+	checkRegionAndOperator(re, tc, co, 1, 0)
+
+	tc = newTestCluster(ctx, opt)
+	co = schedule.NewCoordinator(ctx, tc.RaftCluster, hbStreams)
+
+	re.NoError(tc.addRegionStore(4, 4))
+	re.NoError(tc.addRegionStore(3, 3))
+	re.NoError(tc.addRegionStore(2, 2))
+	re.NoError(tc.addRegionStore(1, 1))
+	re.NoError(tc.putRegion(r))
+	checkRegionAndOperator(re, tc, co, 1, 0)
+	r = r.Clone(core.WithPendingPeers(nil))
+	re.NoError(tc.putRegion(r))
+	checkRegionAndOperator(re, tc, co, 1, 1)
+	op := co.GetOperatorController().GetOperator(1)
+	re.Equal(1, op.Len())
+	re.Equal(uint64(1), op.Step(0).(operator.PromoteLearner).ToStore)
+	checkRegionAndOperator(re, tc, co, 1, 0)
+}
+
+func TestCheckRegionWithScheduleDeny(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+
+	re.NoError(tc.addRegionStore(4, 4))
+	re.NoError(tc.addRegionStore(3, 3))
+	re.NoError(tc.addRegionStore(2, 2))
+	re.NoError(tc.addRegionStore(1, 1))
+	re.NoError(tc.addLeaderRegion(1, 2, 3))
+	region := tc.GetRegion(1)
+	re.NotNil(region)
+	// test with label schedule=deny
+	labelerManager := tc.GetRegionLabeler()
+	labelerManager.SetLabelRule(&labeler.LabelRule{
+		ID:       "schedulelabel",
+		Labels:   []labeler.RegionLabel{{Key: "schedule", Value: "deny"}},
+		RuleType: labeler.KeyRange,
+		Data:     []interface{}{map[string]interface{}{"start_key": "", "end_key": ""}},
+	})
+
+	// should allow to do rule checker
+	re.True(labelerManager.ScheduleDisabled(region))
+	checkRegionAndOperator(re, tc, co, 1, 1)
+
+	// should not allow to merge
+	tc.opt.SetSplitMergeInterval(time.Duration(0))
+	re.NoError(tc.addLeaderRegion(2, 2, 3, 4))
+	re.NoError(tc.addLeaderRegion(3, 2, 3, 4))
+	region = tc.GetRegion(2)
+	re.True(labelerManager.ScheduleDisabled(region))
+	checkRegionAndOperator(re, tc, co, 2, 0)
+	// delete label rule, should allow to do merge
+	labelerManager.DeleteLabelRule("schedulelabel")
+	re.False(labelerManager.ScheduleDisabled(region))
+	checkRegionAndOperator(re, tc, co, 2, 2)
+}
+
+func TestCheckerIsBusy(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(func(cfg *config.ScheduleConfig) {
+		cfg.ReplicaScheduleLimit = 0 // ensure replica checker is busy
+		cfg.MergeScheduleLimit = 10
+	}, nil, nil, re)
+	defer cleanup()
+
+	re.NoError(tc.addRegionStore(1, 0))
+	num := 1 + typeutil.MaxUint64(tc.opt.GetReplicaScheduleLimit(), tc.opt.GetMergeScheduleLimit())
+	var operatorKinds = []operator.OpKind{
+		operator.OpReplica, operator.OpRegion | operator.OpMerge,
+	}
+	for i, operatorKind := range operatorKinds {
+		for j := uint64(0); j < num; j++ {
+			regionID := j + uint64(i+1)*num
+			re.NoError(tc.addLeaderRegion(regionID, 1))
+			switch operatorKind {
+			case operator.OpReplica:
+				op := newTestOperator(regionID, tc.GetRegion(regionID).GetRegionEpoch(), operatorKind)
+				re.Equal(1, co.GetOperatorController().AddWaitingOperator(op))
+			case operator.OpRegion | operator.OpMerge:
+				if regionID%2 == 1 {
+					ops, err := operator.CreateMergeRegionOperator("merge-region", co.GetCluster(), tc.GetRegion(regionID), tc.GetRegion(regionID-1), operator.OpMerge)
+					re.NoError(err)
+					re.Len(ops, co.GetOperatorController().AddWaitingOperator(ops...))
+				}
+			}
+		}
+	}
+	checkRegionAndOperator(re, tc, co, num, 0)
+}
+
+func TestReplica(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(func(cfg *config.ScheduleConfig) {
+		// Turn off balance.
+		cfg.LeaderScheduleLimit = 0
+		cfg.RegionScheduleLimit = 0
+	}, nil, func(co *schedule.Coordinator) { co.Run() }, re)
+	defer cleanup()
+
+	re.NoError(tc.addRegionStore(1, 1))
+	re.NoError(tc.addRegionStore(2, 2))
+	re.NoError(tc.addRegionStore(3, 3))
+	re.NoError(tc.addRegionStore(4, 4))
+
+	stream := mockhbstream.NewHeartbeatStream()
+
+	// Add peer to store 1.
+	re.NoError(tc.addLeaderRegion(1, 2, 3))
+	region := tc.GetRegion(1)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitAddLearner(re, stream, region, 1)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitPromoteLearner(re, stream, region, 1)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+
+	// Peer in store 3 is down, remove peer in store 3 and add peer to store 4.
+	re.NoError(tc.setStoreDown(3))
+	downPeer := &pdpb.PeerStats{
+		Peer:        region.GetStorePeer(3),
+		DownSeconds: 24 * 60 * 60,
+	}
+	region = region.Clone(
+		core.WithDownPeers(append(region.GetDownPeers(), downPeer)),
+	)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitAddLearner(re, stream, region, 4)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitPromoteLearner(re, stream, region, 4)
+	region = region.Clone(core.WithDownPeers(nil))
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+
+	// Remove peer from store 4.
+	re.NoError(tc.addLeaderRegion(2, 1, 2, 3, 4))
+	region = tc.GetRegion(2)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitRemovePeer(re, stream, region, 4)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+
+	// Remove offline peer directly when it's pending.
+	re.NoError(tc.addLeaderRegion(3, 1, 2, 3))
+	re.NoError(tc.setStoreOffline(3))
+	region = tc.GetRegion(3)
+	region = region.Clone(core.WithPendingPeers([]*metapb.Peer{region.GetStorePeer(3)}))
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+}
+
+func TestCheckCache(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(func(cfg *config.ScheduleConfig) {
+		// Turn off replica scheduling.
+		cfg.ReplicaScheduleLimit = 0
+	}, nil, nil, re)
+	defer cleanup()
+
+	re.NoError(tc.addRegionStore(1, 0))
+	re.NoError(tc.addRegionStore(2, 0))
+	re.NoError(tc.addRegionStore(3, 0))
+
+	// Add a peer with two replicas.
+	re.NoError(tc.addLeaderRegion(1, 2, 3))
+	re.NoError(failpoint.Enable("github.com/tikv/pd/pkg/schedule/break-patrol", `return`))
+
+	// case 1: operator cannot be created due to replica-schedule-limit restriction
+	co.GetWaitGroup().Add(1)
+	co.PatrolRegions()
+	re.Len(co.GetCheckerController().GetWaitingRegions(), 1)
+
+	// cancel the replica-schedule-limit restriction
+	cfg := tc.GetScheduleConfig()
+	cfg.ReplicaScheduleLimit = 10
+	tc.SetScheduleConfig(cfg)
+	co.GetWaitGroup().Add(1)
+	co.PatrolRegions()
+	oc := co.GetOperatorController()
+	re.Len(oc.GetOperators(), 1)
+	re.Empty(co.GetCheckerController().GetWaitingRegions())
+
+	// case 2: operator cannot be created due to store limit restriction
+	oc.RemoveOperator(oc.GetOperator(1))
+	tc.SetStoreLimit(1, storelimit.AddPeer, 0)
+	co.GetWaitGroup().Add(1)
+	co.PatrolRegions()
+	re.Len(co.GetCheckerController().GetWaitingRegions(), 1)
+
+	// cancel the store limit restriction
+	tc.SetStoreLimit(1, storelimit.AddPeer, 10)
+	time.Sleep(time.Second)
+	co.GetWaitGroup().Add(1)
+	co.PatrolRegions()
+	re.Len(oc.GetOperators(), 1)
+	re.Empty(co.GetCheckerController().GetWaitingRegions())
+
+	co.GetWaitGroup().Wait()
+	re.NoError(failpoint.Disable("github.com/tikv/pd/pkg/schedule/break-patrol"))
+}
+
+func TestPeerState(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, func(co *schedule.Coordinator) { co.Run() }, re)
+	defer cleanup()
+
+	// Transfer peer from store 4 to store 1.
+	re.NoError(tc.addRegionStore(1, 10))
+	re.NoError(tc.addRegionStore(2, 10))
+	re.NoError(tc.addRegionStore(3, 10))
+	re.NoError(tc.addRegionStore(4, 40))
+	re.NoError(tc.addLeaderRegion(1, 2, 3, 4))
+
+	stream := mockhbstream.NewHeartbeatStream()
+
+	// Wait for schedule.
+	waitOperator(re, co, 1)
+	operatorutil.CheckTransferPeer(re, co.GetOperatorController().GetOperator(1), operator.OpKind(0), 4, 1)
+
+	region := tc.GetRegion(1).Clone()
+
+	// Add new peer.
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitAddLearner(re, stream, region, 1)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitPromoteLearner(re, stream, region, 1)
+
+	// If the new peer is pending, the operator will not finish.
+	region = region.Clone(core.WithPendingPeers(append(region.GetPendingPeers(), region.GetStorePeer(1))))
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+	re.NotNil(co.GetOperatorController().GetOperator(region.GetID()))
+
+	// The new peer is not pending now, the operator will finish.
+	// And we will proceed to remove peer in store 4.
+	region = region.Clone(core.WithPendingPeers(nil))
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitRemovePeer(re, stream, region, 4)
+	re.NoError(tc.addLeaderRegion(1, 1, 2, 3))
+	region = tc.GetRegion(1).Clone()
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitNoResponse(re, stream)
+}
+
+func TestShouldRun(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	tc.RaftCluster.coordinator = co
+	defer cleanup()
+
+	re.NoError(tc.addLeaderStore(1, 5))
+	re.NoError(tc.addLeaderStore(2, 2))
+	re.NoError(tc.addLeaderStore(3, 0))
+	re.NoError(tc.addLeaderStore(4, 0))
+	re.NoError(tc.LoadRegion(1, 1, 2, 3))
+	re.NoError(tc.LoadRegion(2, 1, 2, 3))
+	re.NoError(tc.LoadRegion(3, 1, 2, 3))
+	re.NoError(tc.LoadRegion(4, 1, 2, 3))
+	re.NoError(tc.LoadRegion(5, 1, 2, 3))
+	re.NoError(tc.LoadRegion(6, 2, 1, 4))
+	re.NoError(tc.LoadRegion(7, 2, 1, 4))
+	re.False(co.ShouldRun())
+	re.Equal(2, tc.GetStoreRegionCount(4))
+
+	testCases := []struct {
+		regionID  uint64
+		ShouldRun bool
+	}{
+		{1, false},
+		{2, false},
+		{3, false},
+		{4, false},
+		{5, false},
+		// store4 needs Collect two region
+		{6, false},
+		{7, true},
+	}
+
+	for _, testCase := range testCases {
+		r := tc.GetRegion(testCase.regionID)
+		nr := r.Clone(core.WithLeader(r.GetPeers()[0]))
+		re.NoError(tc.processRegionHeartbeat(nr))
+		re.Equal(testCase.ShouldRun, co.ShouldRun())
+	}
+	nr := &metapb.Region{Id: 6, Peers: []*metapb.Peer{}}
+	newRegion := core.NewRegionInfo(nr, nil)
+	re.Error(tc.processRegionHeartbeat(newRegion))
+	re.Equal(7, co.GetPrepareChecker().GetSum())
+}
+
+func TestShouldRunWithNonLeaderRegions(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	tc.RaftCluster.coordinator = co
+	defer cleanup()
+
+	re.NoError(tc.addLeaderStore(1, 10))
+	re.NoError(tc.addLeaderStore(2, 0))
+	re.NoError(tc.addLeaderStore(3, 0))
+	for i := 0; i < 10; i++ {
+		re.NoError(tc.LoadRegion(uint64(i+1), 1, 2, 3))
+	}
+	re.False(co.ShouldRun())
+	re.Equal(10, tc.GetStoreRegionCount(1))
+
+	testCases := []struct {
+		regionID  uint64
+		ShouldRun bool
+	}{
+		{1, false},
+		{2, false},
+		{3, false},
+		{4, false},
+		{5, false},
+		{6, false},
+		{7, false},
+		{8, false},
+		{9, true},
+	}
+
+	for _, testCase := range testCases {
+		r := tc.GetRegion(testCase.regionID)
+		nr := r.Clone(core.WithLeader(r.GetPeers()[0]))
+		re.NoError(tc.processRegionHeartbeat(nr))
+		re.Equal(testCase.ShouldRun, co.ShouldRun())
+	}
+	nr := &metapb.Region{Id: 9, Peers: []*metapb.Peer{}}
+	newRegion := core.NewRegionInfo(nr, nil)
+	re.Error(tc.processRegionHeartbeat(newRegion))
+	re.Equal(9, co.GetPrepareChecker().GetSum())
+
+	// Now, after server is prepared, there exist some regions with no leader.
+	re.Equal(uint64(0), tc.GetRegion(10).GetLeader().GetStoreId())
+}
+
+func TestAddScheduler(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, func(co *schedule.Coordinator) { co.Run() }, re)
+	defer cleanup()
+
+	re.Len(co.GetSchedulers(), len(config.DefaultSchedulers))
+	re.NoError(co.RemoveScheduler(schedulers.BalanceLeaderName))
+	re.NoError(co.RemoveScheduler(schedulers.BalanceRegionName))
+	re.NoError(co.RemoveScheduler(schedulers.HotRegionName))
+	re.NoError(co.RemoveScheduler(schedulers.BalanceWitnessName))
+	re.NoError(co.RemoveScheduler(schedulers.TransferWitnessLeaderName))
+	re.Empty(co.GetSchedulers())
+
+	stream := mockhbstream.NewHeartbeatStream()
+
+	// Add stores 1,2,3
+	re.NoError(tc.addLeaderStore(1, 1))
+	re.NoError(tc.addLeaderStore(2, 1))
+	re.NoError(tc.addLeaderStore(3, 1))
+	// Add regions 1 with leader in store 1 and followers in stores 2,3
+	re.NoError(tc.addLeaderRegion(1, 1, 2, 3))
+	// Add regions 2 with leader in store 2 and followers in stores 1,3
+	re.NoError(tc.addLeaderRegion(2, 2, 1, 3))
+	// Add regions 3 with leader in store 3 and followers in stores 1,2
+	re.NoError(tc.addLeaderRegion(3, 3, 1, 2))
+
+	oc := co.GetOperatorController()
+
+	// test ConfigJSONDecoder create
+	bl, err := schedulers.CreateScheduler(schedulers.BalanceLeaderType, oc, storage.NewStorageWithMemoryBackend(), schedulers.ConfigJSONDecoder([]byte("{}")))
+	re.NoError(err)
+	conf, err := bl.EncodeConfig()
+	re.NoError(err)
+	data := make(map[string]interface{})
+	err = json.Unmarshal(conf, &data)
+	re.NoError(err)
+	batch := data["batch"].(float64)
+	re.Equal(4, int(batch))
+
+	gls, err := schedulers.CreateScheduler(schedulers.GrantLeaderType, oc, storage.NewStorageWithMemoryBackend(), schedulers.ConfigSliceDecoder(schedulers.GrantLeaderType, []string{"0"}))
+	re.NoError(err)
+	re.NotNil(co.AddScheduler(gls))
+	re.NotNil(co.RemoveScheduler(gls.GetName()))
+
+	gls, err = schedulers.CreateScheduler(schedulers.GrantLeaderType, oc, storage.NewStorageWithMemoryBackend(), schedulers.ConfigSliceDecoder(schedulers.GrantLeaderType, []string{"1"}))
+	re.NoError(err)
+	re.NoError(co.AddScheduler(gls))
+
+	hb, err := schedulers.CreateScheduler(schedulers.HotRegionType, oc, storage.NewStorageWithMemoryBackend(), schedulers.ConfigJSONDecoder([]byte("{}")))
+	re.NoError(err)
+	conf, err = hb.EncodeConfig()
+	re.NoError(err)
+	data = make(map[string]interface{})
+	re.NoError(json.Unmarshal(conf, &data))
+	re.Contains(data, "enable-for-tiflash")
+	re.Equal("true", data["enable-for-tiflash"].(string))
+
+	// Transfer all leaders to store 1.
+	waitOperator(re, co, 2)
+	region2 := tc.GetRegion(2)
+	re.NoError(dispatchHeartbeat(co, region2, stream))
+	region2 = waitTransferLeader(re, stream, region2, 1)
+	re.NoError(dispatchHeartbeat(co, region2, stream))
+	waitNoResponse(re, stream)
+
+	waitOperator(re, co, 3)
+	region3 := tc.GetRegion(3)
+	re.NoError(dispatchHeartbeat(co, region3, stream))
+	region3 = waitTransferLeader(re, stream, region3, 1)
+	re.NoError(dispatchHeartbeat(co, region3, stream))
+	waitNoResponse(re, stream)
+}
+
+func TestPersistScheduler(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	tc, co, cleanup := prepare(nil, nil, func(co *schedule.Coordinator) { co.Run() }, re)
+	hbStreams := co.GetHeartbeatStreams()
+	defer cleanup()
+	defaultCount := len(config.DefaultSchedulers)
+	// Add stores 1,2
+	re.NoError(tc.addLeaderStore(1, 1))
+	re.NoError(tc.addLeaderStore(2, 1))
+
+	re.Len(co.GetSchedulers(), defaultCount)
+	oc := co.GetOperatorController()
+	storage := tc.RaftCluster.storage
+
+	gls1, err := schedulers.CreateScheduler(schedulers.GrantLeaderType, oc, storage, schedulers.ConfigSliceDecoder(schedulers.GrantLeaderType, []string{"1"}))
+	re.NoError(err)
+	re.NoError(co.AddScheduler(gls1, "1"))
+	evict, err := schedulers.CreateScheduler(schedulers.EvictLeaderType, oc, storage, schedulers.ConfigSliceDecoder(schedulers.EvictLeaderType, []string{"2"}))
+	re.NoError(err)
+	re.NoError(co.AddScheduler(evict, "2"))
+	re.Len(co.GetSchedulers(), defaultCount+2)
+	sches, _, err := storage.LoadAllScheduleConfig()
+	re.NoError(err)
+	re.Len(sches, defaultCount+2)
+
+	// remove 5 schedulers
+	re.NoError(co.RemoveScheduler(schedulers.BalanceLeaderName))
+	re.NoError(co.RemoveScheduler(schedulers.BalanceRegionName))
+	re.NoError(co.RemoveScheduler(schedulers.HotRegionName))
+	re.NoError(co.RemoveScheduler(schedulers.BalanceWitnessName))
+	re.NoError(co.RemoveScheduler(schedulers.TransferWitnessLeaderName))
+	re.Len(co.GetSchedulers(), defaultCount-3)
+	re.NoError(co.GetCluster().GetPersistOptions().Persist(storage))
+	co.Stop()
+	co.GetWaitGroup().Wait()
+	// make a new coordinator for testing
+	// whether the schedulers added or removed in dynamic way are recorded in opt
+	_, newOpt, err := newTestScheduleConfig()
+	re.NoError(err)
+	_, err = schedulers.CreateScheduler(schedulers.ShuffleRegionType, oc, storage, schedulers.ConfigJSONDecoder([]byte("null")))
+	re.NoError(err)
+	// suppose we add a new default enable scheduler
+	config.DefaultSchedulers = append(config.DefaultSchedulers, config.SchedulerConfig{Type: "shuffle-region"})
+	defer func() {
+		config.DefaultSchedulers = config.DefaultSchedulers[:len(config.DefaultSchedulers)-1]
+	}()
+	re.Len(newOpt.GetSchedulers(), defaultCount)
+	re.NoError(newOpt.Reload(storage))
+	// only remains 3 items with independent config.
+	sches, _, err = storage.LoadAllScheduleConfig()
+	re.NoError(err)
+	re.Len(sches, 3)
+
+	// option have 6 items because the default scheduler do not remove.
+	re.Len(newOpt.GetSchedulers(), defaultCount+3)
+	re.NoError(newOpt.Persist(storage))
+	tc.RaftCluster.opt = newOpt
+
+	co = schedule.NewCoordinator(ctx, tc.RaftCluster, hbStreams)
+	co.Run()
+	re.Len(co.GetSchedulers(), 3)
+	co.Stop()
+	co.GetWaitGroup().Wait()
+	// suppose restart PD again
+	_, newOpt, err = newTestScheduleConfig()
+	re.NoError(err)
+	re.NoError(newOpt.Reload(storage))
+	tc.RaftCluster.opt = newOpt
+	co = schedule.NewCoordinator(ctx, tc.RaftCluster, hbStreams)
+	co.Run()
+	re.Len(co.GetSchedulers(), 3)
+	bls, err := schedulers.CreateScheduler(schedulers.BalanceLeaderType, oc, storage, schedulers.ConfigSliceDecoder(schedulers.BalanceLeaderType, []string{"", ""}))
+	re.NoError(err)
+	re.NoError(co.AddScheduler(bls))
+	brs, err := schedulers.CreateScheduler(schedulers.BalanceRegionType, oc, storage, schedulers.ConfigSliceDecoder(schedulers.BalanceRegionType, []string{"", ""}))
+	re.NoError(err)
+	re.NoError(co.AddScheduler(brs))
+	re.Len(co.GetSchedulers(), defaultCount)
+
+	// the scheduler option should contain 6 items
+	// the `hot scheduler` are disabled
+	re.Len(co.GetCluster().GetPersistOptions().GetSchedulers(), defaultCount+3)
+	re.NoError(co.RemoveScheduler(schedulers.GrantLeaderName))
+	// the scheduler that is not enable by default will be completely deleted
+	re.Len(co.GetCluster().GetPersistOptions().GetSchedulers(), defaultCount+2)
+	re.Len(co.GetSchedulers(), 4)
+	re.NoError(co.GetCluster().GetPersistOptions().Persist(co.GetCluster().GetStorage()))
+	co.Stop()
+	co.GetWaitGroup().Wait()
+	_, newOpt, err = newTestScheduleConfig()
+	re.NoError(err)
+	re.NoError(newOpt.Reload(co.GetCluster().GetStorage()))
+	tc.RaftCluster.opt = newOpt
+	co = schedule.NewCoordinator(ctx, tc.RaftCluster, hbStreams)
+
+	co.Run()
+	re.Len(co.GetSchedulers(), defaultCount-1)
+	re.NoError(co.RemoveScheduler(schedulers.EvictLeaderName))
+	re.Len(co.GetSchedulers(), defaultCount-2)
+}
+
+func TestRemoveScheduler(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	tc, co, cleanup := prepare(func(cfg *config.ScheduleConfig) {
+		cfg.ReplicaScheduleLimit = 0
+	}, nil, func(co *schedule.Coordinator) { co.Run() }, re)
+	hbStreams := co.GetHeartbeatStreams()
+	defer cleanup()
+
+	// Add stores 1,2
+	re.NoError(tc.addLeaderStore(1, 1))
+	re.NoError(tc.addLeaderStore(2, 1))
+	defaultCount := len(config.DefaultSchedulers)
+
+	re.Len(co.GetSchedulers(), defaultCount)
+	oc := co.GetOperatorController()
+	storage := tc.RaftCluster.storage
+
+	gls1, err := schedulers.CreateScheduler(schedulers.GrantLeaderType, oc, storage, schedulers.ConfigSliceDecoder(schedulers.GrantLeaderType, []string{"1"}))
+	re.NoError(err)
+	re.NoError(co.AddScheduler(gls1, "1"))
+	re.Len(co.GetSchedulers(), defaultCount+1)
+	sches, _, err := storage.LoadAllScheduleConfig()
+	re.NoError(err)
+	re.Len(sches, defaultCount+1)
+
+	// remove all schedulers
+	re.NoError(co.RemoveScheduler(schedulers.BalanceLeaderName))
+	re.NoError(co.RemoveScheduler(schedulers.BalanceRegionName))
+	re.NoError(co.RemoveScheduler(schedulers.HotRegionName))
+	re.NoError(co.RemoveScheduler(schedulers.GrantLeaderName))
+	re.NoError(co.RemoveScheduler(schedulers.BalanceWitnessName))
+	re.NoError(co.RemoveScheduler(schedulers.TransferWitnessLeaderName))
+	// all removed
+	sches, _, err = storage.LoadAllScheduleConfig()
+	re.NoError(err)
+	re.Empty(sches)
+	re.Empty(co.GetSchedulers())
+	re.NoError(co.GetCluster().GetPersistOptions().Persist(co.GetCluster().GetStorage()))
+	co.Stop()
+	co.GetWaitGroup().Wait()
+
+	// suppose restart PD again
+	_, newOpt, err := newTestScheduleConfig()
+	re.NoError(err)
+	re.NoError(newOpt.Reload(tc.storage))
+	tc.RaftCluster.opt = newOpt
+	co = schedule.NewCoordinator(ctx, tc.RaftCluster, hbStreams)
+	co.Run()
+	re.Empty(co.GetSchedulers())
+	// the option remains default scheduler
+	re.Len(co.GetCluster().GetPersistOptions().GetSchedulers(), defaultCount)
+	co.Stop()
+	co.GetWaitGroup().Wait()
+}
+
+func TestRestart(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	tc, co, cleanup := prepare(func(cfg *config.ScheduleConfig) {
+		// Turn off balance, we test add replica only.
+		cfg.LeaderScheduleLimit = 0
+		cfg.RegionScheduleLimit = 0
+	}, nil, func(co *schedule.Coordinator) { co.Run() }, re)
+	hbStreams := co.GetHeartbeatStreams()
+	defer cleanup()
+
+	// Add 3 stores (1, 2, 3) and a region with 1 replica on store 1.
+	re.NoError(tc.addRegionStore(1, 1))
+	re.NoError(tc.addRegionStore(2, 2))
+	re.NoError(tc.addRegionStore(3, 3))
+	re.NoError(tc.addLeaderRegion(1, 1))
+	region := tc.GetRegion(1)
+	co.GetPrepareChecker().Collect(region)
+
+	// Add 1 replica on store 2.
+	stream := mockhbstream.NewHeartbeatStream()
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitAddLearner(re, stream, region, 2)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitPromoteLearner(re, stream, region, 2)
+	co.Stop()
+	co.GetWaitGroup().Wait()
+
+	// Recreate coordinator then add another replica on store 3.
+	co = schedule.NewCoordinator(ctx, tc.RaftCluster, hbStreams)
+	co.GetPrepareChecker().Collect(region)
+	co.Run()
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	region = waitAddLearner(re, stream, region, 3)
+	re.NoError(dispatchHeartbeat(co, region, stream))
+	waitPromoteLearner(re, stream, region, 3)
+}
+
+func TestPauseScheduler(t *testing.T) {
+	re := require.New(t)
+
+	_, co, cleanup := prepare(nil, nil, func(co *schedule.Coordinator) { co.Run() }, re)
+	defer cleanup()
+	_, err := co.IsSchedulerAllowed("test")
+	re.Error(err)
+	co.PauseOrResumeScheduler(schedulers.BalanceLeaderName, 60)
+	paused, _ := co.IsSchedulerPaused(schedulers.BalanceLeaderName)
+	re.True(paused)
+	pausedAt, err := co.GetPausedSchedulerDelayAt(schedulers.BalanceLeaderName)
+	re.NoError(err)
+	resumeAt, err := co.GetPausedSchedulerDelayUntil(schedulers.BalanceLeaderName)
+	re.NoError(err)
+	re.Equal(int64(60), resumeAt-pausedAt)
+	allowed, _ := co.IsSchedulerAllowed(schedulers.BalanceLeaderName)
+	re.False(allowed)
+}
+
+func BenchmarkPatrolRegion(b *testing.B) {
+	re := require.New(b)
+
+	mergeLimit := uint64(4100)
+	regionNum := 10000
+
+	tc, co, cleanup := prepare(func(cfg *config.ScheduleConfig) {
+		cfg.MergeScheduleLimit = mergeLimit
+	}, nil, nil, re)
+	defer cleanup()
+
+	tc.opt.SetSplitMergeInterval(time.Duration(0))
+	for i := 1; i < 4; i++ {
+		if err := tc.addRegionStore(uint64(i), regionNum, 96); err != nil {
+			return
+		}
+	}
+	for i := 0; i < regionNum; i++ {
+		if err := tc.addLeaderRegion(uint64(i), 1, 2, 3); err != nil {
+			return
+		}
+	}
+
+	listen := make(chan int)
+	go func() {
+		oc := co.GetOperatorController()
+		listen <- 0
+		for {
+			if oc.OperatorCount(operator.OpMerge) == mergeLimit {
+				co.Stop()
+				return
+			}
+		}
+	}()
+	<-listen
+
+	co.GetWaitGroup().Add(1)
+	b.ResetTimer()
+	co.PatrolRegions()
+}
+
+func waitOperator(re *require.Assertions, co *schedule.Coordinator, regionID uint64) {
+	testutil.Eventually(re, func() bool {
+		return co.GetOperatorController().GetOperator(regionID) != nil
+	})
+}
+
+func TestOperatorCount(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+	oc := co.GetOperatorController()
+	re.Equal(uint64(0), oc.OperatorCount(operator.OpLeader))
+	re.Equal(uint64(0), oc.OperatorCount(operator.OpRegion))
+
+	re.NoError(tc.addLeaderRegion(1, 1))
+	re.NoError(tc.addLeaderRegion(2, 2))
+	{
+		op1 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpLeader)
+		oc.AddWaitingOperator(op1)
+		re.Equal(uint64(1), oc.OperatorCount(operator.OpLeader)) // 1:leader
+		op2 := newTestOperator(2, tc.GetRegion(2).GetRegionEpoch(), operator.OpLeader)
+		oc.AddWaitingOperator(op2)
+		re.Equal(uint64(2), oc.OperatorCount(operator.OpLeader)) // 1:leader, 2:leader
+		re.True(oc.RemoveOperator(op1))
+		re.Equal(uint64(1), oc.OperatorCount(operator.OpLeader)) // 2:leader
+	}
+
+	{
+		op1 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpRegion)
+		oc.AddWaitingOperator(op1)
+		re.Equal(uint64(1), oc.OperatorCount(operator.OpRegion)) // 1:region 2:leader
+		re.Equal(uint64(1), oc.OperatorCount(operator.OpLeader))
+		op2 := newTestOperator(2, tc.GetRegion(2).GetRegionEpoch(), operator.OpRegion)
+		op2.SetPriorityLevel(constant.High)
+		oc.AddWaitingOperator(op2)
+		re.Equal(uint64(2), oc.OperatorCount(operator.OpRegion)) // 1:region 2:region
+		re.Equal(uint64(0), oc.OperatorCount(operator.OpLeader))
+	}
+}
+
+func TestStoreOverloaded(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+	oc := co.GetOperatorController()
+	lb, err := schedulers.CreateScheduler(schedulers.BalanceRegionType, oc, tc.storage, schedulers.ConfigSliceDecoder(schedulers.BalanceRegionType, []string{"", ""}))
+	re.NoError(err)
+	opt := tc.GetOpts()
+	re.NoError(tc.addRegionStore(4, 100))
+	re.NoError(tc.addRegionStore(3, 100))
+	re.NoError(tc.addRegionStore(2, 100))
+	re.NoError(tc.addRegionStore(1, 10))
+	re.NoError(tc.addLeaderRegion(1, 2, 3, 4))
+	region := tc.GetRegion(1).Clone(core.SetApproximateSize(60))
+	tc.putRegion(region)
+	start := time.Now()
+	{
+		ops, _ := lb.Schedule(tc, false /* dryRun */)
+		re.Len(ops, 1)
+		op1 := ops[0]
+		re.NotNil(op1)
+		re.True(oc.AddOperator(op1))
+		re.True(oc.RemoveOperator(op1))
+	}
+	for {
+		time.Sleep(time.Millisecond * 10)
+		ops, _ := lb.Schedule(tc, false /* dryRun */)
+		if time.Since(start) > time.Second {
+			break
+		}
+		re.Empty(ops)
+	}
+
+	// reset all stores' limit
+	// scheduling one time needs 1/10 seconds
+	opt.SetAllStoresLimit(storelimit.AddPeer, 600)
+	opt.SetAllStoresLimit(storelimit.RemovePeer, 600)
+	time.Sleep(time.Second)
+	for i := 0; i < 10; i++ {
+		ops, _ := lb.Schedule(tc, false /* dryRun */)
+		re.Len(ops, 1)
+		op := ops[0]
+		re.True(oc.AddOperator(op))
+		re.True(oc.RemoveOperator(op))
+	}
+	// sleep 1 seconds to make sure that the token is filled up
+	time.Sleep(time.Second)
+	for i := 0; i < 100; i++ {
+		ops, _ := lb.Schedule(tc, false /* dryRun */)
+		re.Greater(len(ops), 0)
+	}
+}
+
+func TestStoreOverloadedWithReplace(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+	oc := co.GetOperatorController()
+	lb, err := schedulers.CreateScheduler(schedulers.BalanceRegionType, oc, tc.storage, schedulers.ConfigSliceDecoder(schedulers.BalanceRegionType, []string{"", ""}))
+	re.NoError(err)
+
+	re.NoError(tc.addRegionStore(4, 100))
+	re.NoError(tc.addRegionStore(3, 100))
+	re.NoError(tc.addRegionStore(2, 100))
+	re.NoError(tc.addRegionStore(1, 10))
+	re.NoError(tc.addLeaderRegion(1, 2, 3, 4))
+	re.NoError(tc.addLeaderRegion(2, 1, 3, 4))
+	region := tc.GetRegion(1).Clone(core.SetApproximateSize(60))
+	tc.putRegion(region)
+	region = tc.GetRegion(2).Clone(core.SetApproximateSize(60))
+	tc.putRegion(region)
+	op1 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpRegion, operator.AddPeer{ToStore: 1, PeerID: 1})
+	re.True(oc.AddOperator(op1))
+	op2 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpRegion, operator.AddPeer{ToStore: 2, PeerID: 2})
+	op2.SetPriorityLevel(constant.High)
+	re.True(oc.AddOperator(op2))
+	op3 := newTestOperator(1, tc.GetRegion(2).GetRegionEpoch(), operator.OpRegion, operator.AddPeer{ToStore: 1, PeerID: 3})
+	re.False(oc.AddOperator(op3))
+	ops, _ := lb.Schedule(tc, false /* dryRun */)
+	re.Empty(ops)
+	// sleep 2 seconds to make sure that token is filled up
+	time.Sleep(2 * time.Second)
+	ops, _ = lb.Schedule(tc, false /* dryRun */)
+	re.Greater(len(ops), 0)
+}
+
+func TestDownStoreLimit(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+	oc := co.GetOperatorController()
+	rc := co.GetCheckerController().GetRuleChecker()
+
+	tc.addRegionStore(1, 100)
+	tc.addRegionStore(2, 100)
+	tc.addRegionStore(3, 100)
+	tc.addLeaderRegion(1, 1, 2, 3)
+
+	region := tc.GetRegion(1)
+	tc.setStoreDown(1)
+	tc.SetStoreLimit(1, storelimit.RemovePeer, 1)
+
+	region = region.Clone(core.WithDownPeers([]*pdpb.PeerStats{
+		{
+			Peer:        region.GetStorePeer(1),
+			DownSeconds: 24 * 60 * 60,
+		},
+	}), core.SetApproximateSize(1))
+	tc.putRegion(region)
+	for i := uint64(1); i < 20; i++ {
+		tc.addRegionStore(i+3, 100)
+		op := rc.Check(region)
+		re.NotNil(op)
+		re.True(oc.AddOperator(op))
+		oc.RemoveOperator(op)
+	}
+
+	region = region.Clone(core.SetApproximateSize(100))
+	tc.putRegion(region)
+	for i := uint64(20); i < 25; i++ {
+		tc.addRegionStore(i+3, 100)
+		op := rc.Check(region)
+		re.NotNil(op)
+		re.True(oc.AddOperator(op))
+		oc.RemoveOperator(op)
+	}
+}
+
+// FIXME: remove after move into schedulers package
+type mockLimitScheduler struct {
+	schedulers.Scheduler
+	limit   uint64
+	counter *operator.Controller
+	kind    operator.OpKind
+}
+
+func (s *mockLimitScheduler) IsScheduleAllowed(cluster sche.ClusterInformer) bool {
+	return s.counter.OperatorCount(s.kind) < s.limit
+}
+
+func TestController(t *testing.T) {
+	re := require.New(t)
+
+	tc, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+	oc := co.GetOperatorController()
+
+	re.NoError(tc.addLeaderRegion(1, 1))
+	re.NoError(tc.addLeaderRegion(2, 2))
+	scheduler, err := schedulers.CreateScheduler(schedulers.BalanceLeaderType, oc, storage.NewStorageWithMemoryBackend(), schedulers.ConfigSliceDecoder(schedulers.BalanceLeaderType, []string{"", ""}))
+	re.NoError(err)
+	lb := &mockLimitScheduler{
+		Scheduler: scheduler,
+		counter:   oc,
+		kind:      operator.OpLeader,
+	}
+
+	sc := schedule.NewScheduleController(co, lb)
+
+	for i := schedulers.MinScheduleInterval; sc.GetInterval() != schedulers.MaxScheduleInterval; i = sc.GetNextInterval(i) {
+		re.Equal(i, sc.GetInterval())
+		re.Empty(sc.Schedule(false))
+	}
+	// limit = 2
+	lb.limit = 2
+	// count = 0
+	{
+		re.True(sc.AllowSchedule(false))
+		op1 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpLeader)
+		re.Equal(1, oc.AddWaitingOperator(op1))
+		// count = 1
+		re.True(sc.AllowSchedule(false))
+		op2 := newTestOperator(2, tc.GetRegion(2).GetRegionEpoch(), operator.OpLeader)
+		re.Equal(1, oc.AddWaitingOperator(op2))
+		// count = 2
+		re.False(sc.AllowSchedule(false))
+		re.True(oc.RemoveOperator(op1))
+		// count = 1
+		re.True(sc.AllowSchedule(false))
+	}
+
+	op11 := newTestOperator(1, tc.GetRegion(1).GetRegionEpoch(), operator.OpLeader)
+	// add a PriorityKind operator will remove old operator
+	{
+		op3 := newTestOperator(2, tc.GetRegion(2).GetRegionEpoch(), operator.OpHotRegion)
+		op3.SetPriorityLevel(constant.High)
+		re.Equal(1, oc.AddWaitingOperator(op11))
+		re.False(sc.AllowSchedule(false))
+		re.Equal(1, oc.AddWaitingOperator(op3))
+		re.True(sc.AllowSchedule(false))
+		re.True(oc.RemoveOperator(op3))
+	}
+
+	// add a admin operator will remove old operator
+	{
+		op2 := newTestOperator(2, tc.GetRegion(2).GetRegionEpoch(), operator.OpLeader)
+		re.Equal(1, oc.AddWaitingOperator(op2))
+		re.False(sc.AllowSchedule(false))
+		op4 := newTestOperator(2, tc.GetRegion(2).GetRegionEpoch(), operator.OpAdmin)
+		op4.SetPriorityLevel(constant.High)
+		re.Equal(1, oc.AddWaitingOperator(op4))
+		re.True(sc.AllowSchedule(false))
+		re.True(oc.RemoveOperator(op4))
+	}
+
+	// test wrong region id.
+	{
+		op5 := newTestOperator(3, &metapb.RegionEpoch{}, operator.OpHotRegion)
+		re.Equal(0, oc.AddWaitingOperator(op5))
+	}
+
+	// test wrong region epoch.
+	re.True(oc.RemoveOperator(op11))
+	epoch := &metapb.RegionEpoch{
+		Version: tc.GetRegion(1).GetRegionEpoch().GetVersion() + 1,
+		ConfVer: tc.GetRegion(1).GetRegionEpoch().GetConfVer(),
+	}
+	{
+		op6 := newTestOperator(1, epoch, operator.OpLeader)
+		re.Equal(0, oc.AddWaitingOperator(op6))
+	}
+	epoch.Version--
+	{
+		op6 := newTestOperator(1, epoch, operator.OpLeader)
+		re.Equal(1, oc.AddWaitingOperator(op6))
+		re.True(oc.RemoveOperator(op6))
+	}
+}
+
+func TestInterval(t *testing.T) {
+	re := require.New(t)
+
+	_, co, cleanup := prepare(nil, nil, nil, re)
+	defer cleanup()
+
+	lb, err := schedulers.CreateScheduler(schedulers.BalanceLeaderType, co.GetOperatorController(), storage.NewStorageWithMemoryBackend(), schedulers.ConfigSliceDecoder(schedulers.BalanceLeaderType, []string{"", ""}))
+	re.NoError(err)
+	sc := schedule.NewScheduleController(co, lb)
+
+	// If no operator for x seconds, the next check should be in x/2 seconds.
+	idleSeconds := []int{5, 10, 20, 30, 60}
+	for _, n := range idleSeconds {
+		sc.SetInterval(schedulers.MinScheduleInterval)
+		for totalSleep := time.Duration(0); totalSleep <= time.Second*time.Duration(n); totalSleep += sc.GetInterval() {
+			re.Empty(sc.Schedule(false))
+		}
+		re.Less(sc.GetInterval(), time.Second*time.Duration(n/2))
+	}
+}
+
+func waitAddLearner(re *require.Assertions, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+	var res *pdpb.RegionHeartbeatResponse
+	testutil.Eventually(re, func() bool {
+		if res = stream.Recv(); res != nil {
+			return res.GetRegionId() == region.GetID() &&
+				res.GetChangePeer().GetChangeType() == eraftpb.ConfChangeType_AddLearnerNode &&
+				res.GetChangePeer().GetPeer().GetStoreId() == storeID
+		}
+		return false
+	})
+	return region.Clone(
+		core.WithAddPeer(res.GetChangePeer().GetPeer()),
+		core.WithIncConfVer(),
+	)
+}
+
+func waitPromoteLearner(re *require.Assertions, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+	var res *pdpb.RegionHeartbeatResponse
+	testutil.Eventually(re, func() bool {
+		if res = stream.Recv(); res != nil {
+			return res.GetRegionId() == region.GetID() &&
+				res.GetChangePeer().GetChangeType() == eraftpb.ConfChangeType_AddNode &&
+				res.GetChangePeer().GetPeer().GetStoreId() == storeID
+		}
+		return false
+	})
+	// Remove learner than add voter.
+	return region.Clone(
+		core.WithRemoveStorePeer(storeID),
+		core.WithAddPeer(res.GetChangePeer().GetPeer()),
+	)
+}
+
+func waitRemovePeer(re *require.Assertions, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+	var res *pdpb.RegionHeartbeatResponse
+	testutil.Eventually(re, func() bool {
+		if res = stream.Recv(); res != nil {
+			return res.GetRegionId() == region.GetID() &&
+				res.GetChangePeer().GetChangeType() == eraftpb.ConfChangeType_RemoveNode &&
+				res.GetChangePeer().GetPeer().GetStoreId() == storeID
+		}
+		return false
+	})
+	return region.Clone(
+		core.WithRemoveStorePeer(storeID),
+		core.WithIncConfVer(),
+	)
+}
+
+func waitTransferLeader(re *require.Assertions, stream mockhbstream.HeartbeatStream, region *core.RegionInfo, storeID uint64) *core.RegionInfo {
+	var res *pdpb.RegionHeartbeatResponse
+	testutil.Eventually(re, func() bool {
+		if res = stream.Recv(); res != nil {
+			if res.GetRegionId() == region.GetID() {
+				for _, peer := range append(res.GetTransferLeader().GetPeers(), res.GetTransferLeader().GetPeer()) {
+					if peer.GetStoreId() == storeID {
+						return true
+					}
+				}
+			}
+		}
+		return false
+	})
+	return region.Clone(
+		core.WithLeader(region.GetStorePeer(storeID)),
+	)
+}
+
+func waitNoResponse(re *require.Assertions, stream mockhbstream.HeartbeatStream) {
+	testutil.Eventually(re, func() bool {
+		res := stream.Recv()
+		return res == nil
+	})
 }
